+++ crates/hotpath-meta/src/lib_on/streams/wrapper.rs
@@ -22,7 +22,7 @@ pin_project! {
         #[pin]
         inner: S,
         stats_tx: CbSender<StreamEvent>,
-        id: u64,
+        id: u32,
     }
 }
 
@@ -82,7 +82,7 @@ pin_project! {
         #[pin]
         inner: S,
         stats_tx: CbSender<StreamEvent>,
-        id: u64,
+        id: u32,
     }
 }
